web/requirements.txt
Django
psycopg2
djangorestframework==3.0.4
djangorestframework-gis
markdown
django-filter
